refactor: options and p2p module
to seperate native code, and also add code for browser
